moved dry translator to separate file
